Split public proxy config code
